Move UserProfile ViewModels to their own folder.
