+++ Project.Tests/Controllers/AdminControllerTests.cs
@@ -7,7 +7,7 @@ using Project.Core.Account;
 using Project.Tests.Utils;
 using Project.UserProfileDomain.Models;
 using Project.UserProfileDomain.Repositories;
-using Project.ViewModels;
+using Project.ViewModels.UserProfile;
 using Project.ViewModels.Admin;
 using System;
 using System.Collections.Generic;
