+++ Project/ViewModels/UserProfile/GoalVM.cs
@@ -1,4 +1,4 @@
-﻿namespace Project.ViewModels {
+﻿namespace Project.ViewModels.UserProfile {
     public class GoalVM {
         public int Id { get; set; }
