+++ Project/ViewModels/UserProfile/GroupVM.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace Project.ViewModels
-{
+namespace Project.ViewModels.UserProfile {
     public class GroupVM
     {
         public int Id { get; set; }
