+++ Project/ViewModels/UserProfile/InterestVM.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace Project.ViewModels {
+namespace Project.ViewModels.UserProfile {
     public class InterestVM
     {
         public int Id { get; set; }
