+++ Project/ViewModels/UserProfile/UserInterestVM.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Project.ViewModels.UserProfile {
+    public class UserInterestVM {
+
+        public string UserProfileId { get; set; }
+
+        public int InterestId { get; set; }
+
+    }
+}
