+++ Project/ViewModels/UserProfile/UserProfileRefVM.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace Project.ViewModels
-{
+namespace Project.ViewModels.UserProfile {
     public class UserProfileRefVM
     {
         public int Id { get; set; }
